# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Agent Studio API
  version: 0.1.0
  description: 轻量化 Agent Studio 的本地 API。
servers:
  - url: http://localhost:8080
paths:
...
  /api/node-types/{type}/{version}/resolve:
    post:
      operationId: resolveNodeType
      parameters:
        - $ref: "#/components/parameters/NodeType"
        - $ref: "#/components/parameters/NodeVersion"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [config]
              properties:
                config:
                  type: object
                  additionalProperties: true
      responses:
        "200":
          description: 解析后的端口
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolvedPorts"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /api/node-index/status:
    get:
      operationId: getNodeIndexStatus
      responses:
        "200":
          description: 当前本地节点索引状态
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeIndexStatus"
        "500":
          $ref: "#/components/responses/Error"
  /api/node-packages:
    get:
      operationId: listNodePackages
      parameters:
        - name: q
          in: query
          schema:
            type: string
            maxLength: 128
        - name: category
          in: query
          style: form
          explode: true
          schema:
            type: array
            maxItems: 8
            items:
              type: string
              minLength: 1
              maxLength: 32
              pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
        - name: compatible
          in: query
          schema:
            type: boolean
            default: true
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 10000
            default: 0
      responses:
        "200":
          description: 节点包搜索结果
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodePackageSearchResult"
        "400":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/node-package:
    get:
      operationId: getNodePackage
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 512
      responses:
        "200":
          description: 节点包详情
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodePackageDetail"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/workflows:
    get:
      operationId: listWorkflows
      responses:
        "200":
          description: 工作流列表
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowRequest"
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/workflows/{id}:
    parameters:
      - $ref: "#/components/parameters/WorkflowID"
    get:
      operationId: getWorkflow
      responses:
        "200":
          description: 工作流草稿
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "404":
          $ref: "#/components/responses/Error"
    put:
      operationId: saveWorkflowDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SaveDraftRequest"
      responses:
        "200":
          description: 已保存
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/workflows/{id}/template:
    parameters:
      - $ref: "#/components/parameters/WorkflowID"
    get:
      operationId: exportWorkflowTemplate
      parameters:
        - name: draftRevision
          in: query
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
      responses:
        "200":
          description: 规范化工作流模板下载
          headers:
            Content-Disposition:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplate"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/workflows/{id}/validate:
    parameters:
      - $ref: "#/components/parameters/WorkflowID"
    post:
      operationId: validateWorkflow
      responses:
        "200":
          description: 校验结果
          content:
            application/json:
              schema:
                type: object
                required: [valid, issues]
                properties:
                  valid:
                    type: boolean
                  issues:
                    type: array
                    items:
                      $ref: "#/components/schemas/ValidationIssue"
  /api/workflows/{id}/test-runs:
    parameters:
      - $ref: "#/components/parameters/WorkflowID"
    post:
      operationId: runWorkflowDraft
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftRunRequest"
      responses:
        "200":
          description: 每行一个 RunEvent 的 NDJSON 流
          content:
            application/x-ndjson:
              schema:
                type: string
                description: 每行符合 RunEvent Schema。
        "400":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/workflows/{id}/publish:
    parameters:
      - $ref: "#/components/parameters/WorkflowID"
    post:
      operationId: publishWorkflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [draftRevision]
              properties:
                draftRevision:
                  type: integer
                  format: int64
      responses:
        "201":
          description: 不可变工作流版本
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowVersion"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/workflows/{id}/runs:
    parameters:
      - $ref: "#/components/parameters/WorkflowID"
    get:
      operationId: listWorkflowRuns
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        "200":
          description: 运行记录
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Run"
  /api/workflow-templates/preview:
    post:
      operationId: previewWorkflowTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreviewWorkflowTemplateRequest"
      responses:
        "200":
          description: 模板兼容性预览
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowTemplatePreview"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/workflow-templates/import:
    post:
      operationId: importWorkflowTemplate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportWorkflowTemplateRequest"
      responses:
        "201":
          description: 已创建未发布工作流草稿
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
        "500":
          $ref: "#/components/responses/Error"
  /api/agents/{slug}:
    parameters:
      - $ref: "#/components/parameters/AgentSlug"
    get:
      operationId: getAgentManifest
      responses:
        "200":
          description: 当前发布页 Manifest
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentManifest"
        "404":
          $ref: "#/components/responses/Error"
  /api/agents/{slug}/runs:
    parameters:
      - $ref: "#/components/parameters/AgentSlug"
    post:
      operationId: runAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRunRequest"
      responses:
        "200":
          description: 每行一个 RunEvent 的 NDJSON 流
          content:
            application/x-ndjson:
              schema:
                type: string
                description: 每行符合 RunEvent Schema。
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/runs/{id}:
    get:
      operationId: getRun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 运行详情
          content:
            application/json:
              schema:
                type: object
                required: [run, nodeRuns]
                properties:
                  run:
                    $ref: "#/components/schemas/Run"
                  nodeRuns:
                    type: array
                    items:
                      $ref: "#/components/schemas/NodeRun"
        "404":
          $ref: "#/components/responses/Error"
  /api/runs/{id}/debug:
    get:
      operationId: getRunDebug
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 运行回放概览
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebugOverview"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/runs/{id}/events:
    get:
      operationId: listRunEvents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: afterSequence
          in: query
          schema:
            type: integer
            format: int64
            minimum: 0
            default: 0
      responses:
        "200":
          description: 运行事件页
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEventPage"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
  /api/runs/{id}/nodes/{nodeId}/rerun-preview:
    get:
      operationId: previewNodeRerun
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 局部重跑预览
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RerunPreview"
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
  /api/runs/{id}/nodes/{nodeId}/reruns:
    post:
      operationId: rerunFromNode
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: nodeId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RerunRequest"
      responses:
        "200":
          description: 每行一个新 Debug RunEvent 的 NDJSON 流
          content:
            application/x-ndjson:
              schema:
                type: string
                description: 每行符合 RunEvent Schema。
        "400":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
        "422":
          $ref: "#/components/responses/Error"
components:
  parameters:
    WorkflowID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    AgentSlug:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    NodeType:
      name: type
      in: path
      required: true
      schema:
        type: string
    NodeVersion:
      name: version
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: 稳定错误响应
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    PortDefinition:
      type: object
      required: [key, title, type, required, cardinality]
      properties:
        key:
          type: string
        title:
          type: string
        type:
          type: string
          enum: [string, number, boolean, json, any]
        required:
          type: boolean
        cardinality:
          type: string
          enum: [one, single-active]
    ResolvedPorts:
      type: object
      required: [inputs, outputs]
      properties:
        inputs:
          type: array
          items:
            $ref: "#/components/schemas/PortDefinition"
        outputs:
          type: array
          items:
            $ref: "#/components/schemas/PortDefinition"
...
    NodeIndexStatus:
      type: object
      additionalProperties: false
      required: [source, release, generatedAt, packageCount, compatiblePackageCount, runtimeVersion, nodeAPI, stale, warningCode]
      properties:
        source:
          type: string
          enum: [embedded, cache]
        release:
          $ref: "#/components/schemas/NodeIndexStableSemver"
        generatedAt:
          type: string
          format: date-time
          pattern: "^[0-9]{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12][0-9]|3[01])T(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]Z$"
        packageCount:
          type: integer
          minimum: 0
          maximum: 1000
        compatiblePackageCount:
          type: integer
          minimum: 0
          maximum: 1000
        runtimeVersion:
          type: string
          minLength: 1
          maxLength: 128
        nodeAPI:
          type: string
          minLength: 1
          maxLength: 128
        stale:
          type: boolean
        warningCode:
          oneOf:
            - type: string
              enum: [INDEX_EMBEDDED_SNAPSHOT, INDEX_CONTENT_INVALID]
            - type: "null"
    NodePackageSearchResult:
      type: object
      additionalProperties: false
      required: [release, total, offset, limit, items]
      properties:
        release:
          $ref: "#/components/schemas/NodeIndexStableSemver"
        total:
          type: integer
          minimum: 0
          maximum: 1000
        offset:
          type: integer
          minimum: 0
          maximum: 10000
        limit:
          type: integer
          minimum: 1
          maximum: 100
        items:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/IndexedNodePackageSummary"
...
    NodePackageDetail:
      type: object
      additionalProperties: false
      required: [name, categories, keywords, versions, recommendedVersion, reasons, assessments]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 512
        categories:
          $ref: "#/components/schemas/NodePackageCategories"
        keywords:
          $ref: "#/components/schemas/NodePackageKeywords"
        versions:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/IndexedNodePackageVersion"
        recommendedVersion:
          oneOf:
            - $ref: "#/components/schemas/NodePackageVersionSummary"
            - type: "null"
        reasons:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/NodePackageCompatibilityReason"
        assessments:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/NodePackageVersionAssessment"
...
    NodePackageCategories:
      type: array
      maxItems: 8
      uniqueItems: true
      items:
        type: string
        minLength: 1
        maxLength: 32
        pattern: "^[a-z0-9]+(?:-[a-z0-9]+)*$"
    NodePackageKeywords:
      type: array
      maxItems: 16
      uniqueItems: true
      items:
        type: string
        minLength: 1
        maxLength: 64
    NodePackageCompatibilityReason:
      type: string
      enum: [runtime_invalid, node_api_mismatch, runtime_too_old, runtime_too_new, no_active_stable_version]
...
    NodeIndexStableSemver:
      type: string
      minLength: 1
      maxLength: 128
      pattern: "^v(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)\\.(0|[1-9][0-9]*)$"
    ValidationIssue:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        nodeId:
          type: string
        edgeId:
          type: string
        path:
          type: string
        packageName:
          type: string
        packageVersion:
          type: string
    WorkflowTemplate:
      $ref: "#/components/schemas/WorkflowTemplateV1Alpha2"
...
    PreviewWorkflowTemplateRequest:
      type: object
      additionalProperties: false
      required: [template]
      properties:
        template:
          oneOf:
            - $ref: "#/components/schemas/WorkflowTemplateV1Alpha1"
            - $ref: "#/components/schemas/WorkflowTemplateV1Alpha2"
    ImportWorkflowTemplateRequest:
      type: object
      additionalProperties: false
      required: [template, name, slug, description]
      properties:
        template:
          oneOf:
            - $ref: "#/components/schemas/WorkflowTemplateV1Alpha1"
            - $ref: "#/components/schemas/WorkflowTemplateV1Alpha2"
        name:
          type: string
        slug:
          type: string
        description:
          type: string
    WorkflowTemplatePreview:
      type: object
      additionalProperties: false
      required: [valid, metadata, summary, issues]
      properties:
        valid:
          type: boolean
        metadata:
          $ref: "#/components/schemas/WorkflowTemplateMetadata"
        summary:
          $ref: "#/components/schemas/WorkflowTemplateSummary"
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
...
    Workflow:
      type: object
      required: [id, name, slug, description, draftGraph, draftRevision, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        draftGraph:
          $ref: "#/components/schemas/Graph"
        draftRevision:
          type: integer
          format: int64
        publishedVersionId:
          type: [string, "null"]
          format: uuid
        publishedVersion:
          type: [integer, "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkflowVersion:
      type: object
      required: [id, workflowId, version, graph, inputSchema, createdAt]
      properties:
        id:
          type: string
          format: uuid
        workflowId:
          type: string
          format: uuid
        version:
          type: integer
        graph:
          $ref: "#/components/schemas/Graph"
        inputSchema:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
    AgentManifest:
      type: object
      required: [workflowVersionId, version, title, description, inputSchema]
      properties:
        workflowVersionId:
          type: string
          format: uuid
        version:
          type: integer
        title:
          type: string
        description:
          type: string
        inputSchema:
          type: object
          additionalProperties: true
    Run:
      type: object
      required: [id, workflowId, mode, status, input, startedAt]
      properties:
        id:
          type: string
          format: uuid
        workflowId:
          type: string
          format: uuid
        workflowVersionId:
          type: [string, "null"]
          format: uuid
        draftRevision:
          type: [integer, "null"]
          format: int64
        sourceRunId:
          type: [string, "null"]
          format: uuid
        sourceNodeId:
          type: [string, "null"]
        graphSnapshot:
          oneOf:
            - $ref: "#/components/schemas/Graph"
            - type: "null"
        mode:
          type: string
          enum: [test, published, debug]
        status:
          type: string
          enum: [running, completed, failed, cancelled]
        input: {}
        output: {}
        error:
          oneOf:
            - $ref: "#/components/schemas/PublicError"
            - type: "null"
        startedAt:
          type: string
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
    NodeRun:
      type: object
      required: [id, runId, nodeId, nodeType, status]
      properties:
        id:
          type: string
          format: uuid
        runId:
          type: string
          format: uuid
        nodeId:
          type: string
        nodeType:
          type: string
        status:
          type: string
        input: {}
        output: {}
        error:
          oneOf:
            - $ref: "#/components/schemas/PublicError"
            - type: "null"
        startedAt:
          type: [string, "null"]
          format: date-time
        endedAt:
          type: [string, "null"]
          format: date-time
...
    DebugOverview:
      type: object
      required: [run, graph, nodeRuns, sourceChain, replayAvailable, rerunAvailable]
      properties:
        run:
          $ref: "#/components/schemas/Run"
        graph:
          $ref: "#/components/schemas/Graph"
        nodeRuns:
          type: array
          items:
            $ref: "#/components/schemas/NodeRun"
        sourceChain:
          type: array
          items:
            $ref: "#/components/schemas/DebugSource"
        replayAvailable:
          type: boolean
        rerunAvailable:
          type: boolean
        unavailableReason:
          type: string
    RunEventPage:
      type: object
      required: [events, nextAfterSequence]
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/RunEvent"
        nextAfterSequence:
          type: integer
          format: int64
    ExecutionSafety:
      type: string
      enum: [pure, read_only, side_effect]
...
    RerunPreview:
      type: object
      required: [sourceRunId, sourceNodeId, entryInput, entryInputRedactedPaths, activeNodes, frozenEdges, effectiveSafety, requiresConfirmation]
      properties:
        sourceRunId:
          type: string
          format: uuid
        sourceNodeId:
          type: string
        entryInput:
          type: object
          additionalProperties: true
        entryInputRedactedPaths:
          type: array
          items:
            type: string
        activeNodes:
          type: array
          items:
            $ref: "#/components/schemas/RerunNode"
        frozenEdges:
          type: array
          items:
            $ref: "#/components/schemas/FrozenEdgePreview"
        effectiveSafety:
          $ref: "#/components/schemas/ExecutionSafety"
        requiresConfirmation:
          type: boolean
    RerunRequest:
      type: object
      additionalProperties: false
      required: [entryInput, confirmSideEffects]
      properties:
        entryInput:
          type: object
          additionalProperties: true
        confirmSideEffects:
          type: boolean
...
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    CreateWorkflowRequest:
      type: object
      additionalProperties: false
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
    SaveDraftRequest:
      type: object
      additionalProperties: false
      required: [draftRevision, graph]
      properties:
        draftRevision:
          type: integer
          format: int64
        graph:
          $ref: "#/components/schemas/Graph"
    DraftRunRequest:
      type: object
      additionalProperties: false
      required: [draftRevision, input]
      properties:
        draftRevision:
          type: integer
          format: int64
        input:
          type: object
          additionalProperties: true
    AgentRunRequest:
      type: object
      additionalProperties: false
      required: [workflowVersionId, input]
      properties:
        workflowVersionId:
          type: string
          format: uuid
        input:
          type: object
          additionalProperties: true